Move scheduler to background tasks only
