+++ cq/management/commands/cq_runworker.py
@@ -1,37 +1,23 @@
-from django.core.management.base import CommandError
-from django.conf import settings
+import logging
+from threading import Thread
+
 from channels.management.commands.runworker import Command as BaseCommand
+
+from django.conf import settings
+
+logger = logging.getLogger('cq')
+
+
+def launch_scheduler(*args, **kwargs):
+    from ...scheduler import scheduler
+    logger.info('Launching CQ scheduler.')
+    thread = Thread(name='scheduler', target=scheduler)
+    thread.daemon = True
+    thread.start()
 
 
 class Command(BaseCommand):
-    def add_arguments(self, parser):
-        super().add_arguments(parser)
-        parser.add_argument(
-            '--web-threads', action='store',
-            type=int, help='Number of threads to execute.'
-        )
-        parser.add_argument(
-            '--worker-threads', action='store',
-            type=int, help='Number of threads to execute.'
-        )
-
     def handle(self, *args, **options):
-        try:
-            web_threads = int(options.get('web_threads', 0))
-        except TypeError:
-            web_threads = 0
-        try:
-            wkr_threads = int(options.get('worker_threads', 0))
-        except TypeError:
-            wkr_threads = 0
-        layer = getattr(settings, 'CQ_CHANNEL_LAYER', None)
-        if layer:
-            options['layer'] = layer
-        if (web_threads or wkr_threads):
-            options['threads'] = web_threads + wkr_threads
-            if web_threads:
-                if options.get('thread_only_channels', None) is None:
-                    options['thread_only_channels'] = []
-                for ii in range(web_threads):
-                    options['thread_only_channels'].append('%d,http.*' % ii)
+        if getattr(settings, 'CQ_SCHEDULER', True):
+            launch_scheduler()
         super().handle(*args, **options)
